2. Тесты для идеальный вес [#334861]

+++ src/ru/job4j/calculator/Fit.java
@@ -2,18 +2,18 @@ package ru.job4j.calculator;
 
 public class Fit {
 
-    public static double manWeight(short height) {
+    public static double manWeight(double height) {
         double rsl = (height - 100) * 1.15;
         return rsl;
     }
 
-    public static double womanWeight(short height) {
+    public static double womanWeight(double height) {
         double rsl = (height - 110) * 1.15;
         return rsl;
     }
 
     public static void main(String[] args) {
-        short height = 187;
+        double height = 187;
         double man = Fit.manWeight(height);
         double woman = Fit.womanWeight(height);
         System.out.println("Man 187 is " + man);
